feat: complete gym management backend with robust database entities, Flyway migrations, full CRUD endpoints, and comprehensive tests

+++ src/main/java/com/example/demo/dtos/rental/response/MaintenanceLogResponseDto.java
@@ -2,6 +2,7 @@ package com.example.demo.dtos.rental.response;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -9,7 +10,7 @@ public class MaintenanceLogResponseDto {
     private Long id;
     private LocalDate dateSent;
     private LocalDate dateReturned;
-    private Double cost;
+    private BigDecimal cost;
     private String description;
     private EquipmentResponseDto equipment;
 }
